Test reading Xlsx column/row attrs with ReadFilter
When reading an Xlsx file using a ReadFilter it is posible for
`ColumnAndRowAttributes::load()` to (I think) erroneously skip loading
of column or row attributes.

This is because `isFilteredColumn` and `isFilteredRow` will return early
with a true (ie the column or row should be considered filtered and
their attrbutes discarded) at the first false returned from the
ReadFilter. All rows for a column must past the filter for the loop to
return false. I've also created the same test of column widths for row
heights.

 * ColumnWidthTest.php
   Added test with a read filter.
 * RowHeightTest.php
   Mirrored column wdith tests for row height